requirements.txt
amqp==2.5.2
asgiref==3.2.3
beautifulsoup4==4.8.1
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.3
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-filter==2.2.0
django-jalali==3.3.0
django-timezone-field==4.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
jdatetime==3.6.2
kombu==4.6.7
lxml==4.5.0
Markdown==3.2.1
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==3.0.0
